+++ CHANGELOG.md
@@ -5,3 +5,13 @@ Ralph Change Log
 -----------------
 
 - Initial public release
+
+2018-04-06 v0.5
+-----------------
+
+- Updated codebase to use PHP 7 typehints.
+- Migrate tests from Codeception to PHPUnit.
+- Add ResultSet class to make working with returned results easier.
+- Changed namespace from CCR\Sparql to CCR\Ralph.
+- Renamed SparqlClient to Client.
+- Added use of static analysis with PHPStan and improved Makefile.